+++ src/p_doors.c
@@ -246,10 +246,10 @@ void EV_VerticalDoor(line_t *line, mobj_t *thing)
       break;
 
    }
-	
+
    /* if the sector has an active thinker, use it */
    sec = sides[ line->sidenum[side^1]] .sector;
-   secnum = sec-sectors;
+   secnum = (int)(sec-sectors);
    if(sec->specialdata)
    {
       door = sec->specialdata;
